requirements.txt

Flask==2.3.2
Flask-CORS==3.0.10
tensorflow==2.10.0
h5py==3.7.0
numpy==1.23.5
Pillow==9.4.0
gunicorn==20.1.0